requirements.txt

Pillow==7.0.0